requirements.txt
streamlit
pandas
openpyxl
pytesseract
Pillow